Add synchronous service
Creates and Injectable HeroService, hero.service.ts plus mock-heroes.ts, with a getHeroes method to return the
list of heroes.

Also add a private heroService property to AppComponent which is the
target for a HeroService provider.

Changed AppComponent to implement OnInit interface, add ngOnInit, to
initialize the private heroService property.